examples/sumofinance/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
streamlit>=1.31.0
pydantic>=2.9.0
numpy>=2.0.0
pyotp>=2.9.0
PyJWT>=2.8.0
qrcode>=7.4.2
